Norminetted almost all of my code, very good progress. TODO: Need to check on shlvl still and also try to implement 0 expansion
