+++ Makefile
@@ -6,7 +6,7 @@
 #    By: frey-gal <frey-gal@student.42barcelona.co  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/04/20 02:23:03 by frey-gal          #+#    #+#              #
-#    Updated: 2025/06/23 07:16:49 by frey-gal         ###   ########.fr        #
+#    Updated: 2025/07/10 21:27:18 by frey-gal         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,8 +20,10 @@ RM			= rm -f
 
 # ================================== SOURCES ================================== #
 
-SRC			= $(addprefix srcs/minishell_, exec.c init.c signals.c tokens.c \
-				utils.c builtins.c memory.c variables.c processing.c)
+SRC			= $(addprefix srcs/minishell_, builtins_chdir.c builtins_extra.c \
+			  builtins_vars.c env_helpers.c env_init.c exec.c expansion.c \
+			  expansion_helpers.c init.c memory.c signals.c tokens.c \
+			  utils.c)
 
 OBJ			= $(SRC:.c=.o)
