+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: miggarc2 <miggarc2@student.42barcelona.co  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/29 17:10:44 by miggarc2          #+#    #+#             */
-/*   Updated: 2025/06/23 07:21:23 by frey-gal         ###   ########.fr       */
+/*   Updated: 2025/07/10 21:30:26 by frey-gal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,6 +31,15 @@
 
 extern volatile sig_atomic_t	g_signal_code;
 
+typedef struct s_builder
+{
+	char	*new_token;
+	int		i;
+	int		j;
+	bool	in_single;
+	bool	in_double;
+}	t_builder;
+
 typedef struct s_var
 {
 	char			*line;
@@ -53,7 +62,6 @@ typedef struct s_var
 }					t_var;
 
 // minishell_exec.c
-void	ms_exit_func_handle(t_var *var);
 bool	ms_exec_builtins(t_var *var, int i, bool child);
 void	ft_exec_child(t_var *var, int i, int pipes);
 int		ms_pipex(t_var *var);
@@ -89,34 +97,48 @@ void	*memlist_alloc(t_var *var, size_t size);
 void	*memlist_add(t_var *var, void *ptr);
 void	memlist_free_ptr(t_var *var, void *ptr);
 
-// minishell_builtins.c
+// BUILTINS
+
+// minishell_builtins_extra.c
 void	ms_echo(char **tokens);
 char	*getcwd_plus(t_var *var);
 void	ms_pwd(t_var *var);
-int		bad_status(int status);
+void	ms_exit_builtin(t_var *var);
+// minishell_builtins_chdir.c
+//static int	bad_status(int status);
 //static void	cd_home(t_var *var);
 //static void	cd_previous(t_var *var);
+//static void	cd_todir(t_var *var, char **tokens)
 void	ms_cd(t_var *var, char **tokens);
+// minishell_builtins_vars.c
 //static void	export_print(t_var *var);
+//static void	export_add(t_var *var, char *token)
 void	ms_export(t_var *var, char **tokens);
 void	ms_unset(t_var *var, char **tokens);
 void	ms_env(t_var *var);
 
-// minishell_variables.c
+// ENVIRONMENT VARIABLES
+
+// minishell_env_init.c
+void	modify_env_var(t_var *var, char *var_name, char *new_content);
+void	add_shlvl(t_var *var, char *shlvl);
+void	create_env(t_var *var, char **env);
+// minishell_env_helpers.c
 void	varlist_clean(t_var	*var);
 void	add_env_var(t_var *var, char *variable);
 bool	find_env_var(t_var *var, char *var_name);
 void	remove_env_var(t_var *var, char *var_name);
 char	*get_env_var(t_var *var, char *variable);
-void	modify_env_var(t_var *var, char *var_name, char *new_content);
-void	add_shlvl(t_var *var, char *shlvl);
-void	create_env(t_var *var, char **env);
 
-// minishell_processing.c
-int		var_len_diff(t_var *var, char *str);
-char	*var_finder(t_var *var, char *str, char *new_token);
+// VARIABLE EXPANSION
+// minishell_expansion.c
+//static void token_size_logic(t_var *var, char *token, int *len, t_builder *b)
 int		new_token_size(t_var *var, char *token);
+//static void token_builder_logic(t_builder *b, char *token, t_var *var)
 char	*token_builder(t_var *var, char *token);
 bool	expand_cmd(t_var *var);
+// minishell_expansion_helpers.c
+int		var_len_diff(t_var *var, char *str);
+char	*var_finder(t_var *var, char *str, char *new_token);
 
 #endif
